Refactor: Sign In With Apple 코드 ViewModel로 이동
- Sign In With Apple 코드 ViewModel로 이동
	- 명확한 ViewModel의 역할을 위해
