+++ SampleProject/Socially/Socially/AuthViewModel.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import FirebaseAuth
 import AuthenticationServices
 import CryptoKit
+import FirebaseFirestore
 
 class AuthViewModel: ObservableObject {
     
@@ -19,7 +20,65 @@ class AuthViewModel: ObservableObject {
      즉, 인증 과정을 위한 임의의 랜덤 난수
      */
     
-    func randomNonceString(length: Int = 32) -> String { precondition(length > 0)
+    
+    //MARK: - Sign In With Apple Methods
+    func signInWithAppleRequest(request: ASAuthorizationAppleIDRequest) {
+        let nonce = randomNonceString()
+        currentNonce = nonce
+        request.requestedScopes = [.email]
+        request.nonce = sha256(nonce)
+    }
+    
+    func signInWithAppleCompletion(result: Result<ASAuthorization, any Error>) {
+        switch result {
+        case .success(let authResults):
+            switch authResults.credential {
+            case let appleIDCredential as ASAuthorizationAppleIDCredential:
+                guard let nonce = currentNonce else {
+                    fatalError("Invalid state: A login callback was received, but no login request was sent.")
+                }
+                guard let appleIDToken = appleIDCredential.identityToken else {
+                    fatalError("Invalid state: A login callback was received, but no login request was sent.")
+                }
+                guard let idTokenString = String(data: appleIDToken, encoding: .utf8) else { print("Unable to serialize token string from data: \(appleIDToken.debugDescription)")
+                    return
+                }
+                
+                let credential = OAuthProvider.credential(withProviderID: "apple.com",
+                                                          idToken: idTokenString,
+                                                          rawNonce: nonce)
+                
+                Auth.auth().signIn(with: credential) { (authResult, error) in
+                    if let error = error { print(error.localizedDescription)
+                        return
+                    }
+                    print("signed in")
+                    guard let user = authResult?.user else { return }
+                    
+                    
+                    let userData = ["email": user.email, "uid": user.uid]
+                    Firestore.firestore().collection("User")
+                        .document(user.uid)
+                        .setData(userData as [String : Any]) { _ in
+                            print("DEBUG: Did upload user data.")
+                        }
+                }
+                print("\(String(describing: Auth.auth().currentUser?.uid))")
+            default:
+                break
+            }
+        case .failure(let error):
+            print(error.localizedDescription)
+        }
+    }
+    
+    
+    
+    
+    
+    
+    //MARK: - private methods
+    private func randomNonceString(length: Int = 32) -> String { precondition(length > 0)
         let charset: [Character] = Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
         var result = ""
         var remainingLength = length
@@ -49,7 +108,7 @@ class AuthViewModel: ObservableObject {
         return result
     }
     
-    func sha256(_ input: String) -> String {
+    private func sha256(_ input: String) -> String {
         let inputData = Data(input.utf8)
         let hashedData = SHA256.hash(data: inputData)
         let hashString = hashedData.compactMap { String(format: "%02x", $0) }.joined()
